+++ aten/src/ATen/TensorOptions.h
@@ -6,7 +6,6 @@
 #include <ATen/DeviceGuard.h>
 #include <ATen/core/Layout.h>
 #include <ATen/core/ScalarType.h>
-#include <ATen/Tensor.h>
 #include <ATen/Type.h>
 
 #include <cstddef>
@@ -163,6 +162,8 @@ struct AT_API TensorOptions {
   }
 
  private:
+  // WARNING: If you edit TensorOptions to add more options, you
+  // must adjust the implementation of Tensor::options
   ScalarType dtype_{kFloat};
   Device device_{Device::Type::CPU};
   Layout layout_{Layout::Strided};
@@ -200,49 +201,6 @@ inline TensorOptions requires_grad(bool requires_grad = true) {
   return TensorOptions().requires_grad(requires_grad);
 }
 
-/// From Tensor.h
-inline TensorOptions Tensor::options() const {
-  return TensorOptions().dtype(dtype())
-                        .device(device())
-                        .layout(layout())
-                        .is_variable(is_variable());
-}
-
-namespace detail {
-inline Tensor to(
-    const Tensor& tensor,
-    const TensorOptions& options,
-    bool non_blocking) {
-  // Don't copy if the options match.
-  if (tensor.options() == options) {
-    return tensor;
-  }
-  DeviceGuard guard(options.device());
-  return options.type().copy(tensor, non_blocking);
-}
-} // namespace detail
-
-inline Tensor Tensor::to(Device device, ScalarType dtype, bool non_blocking)
-    const {
-  if (this->device() == device && this->dtype() == dtype) {
-    return *this;
-  }
-  return detail::to(*this, options().device(device).dtype(dtype), non_blocking);
-}
-
-inline Tensor Tensor::to(ScalarType dtype, bool non_blocking) const {
-  if (this->dtype() == dtype) {
-    return *this;
-  }
-  return detail::to(*this, options().dtype(dtype), non_blocking);
-}
-
-inline Tensor Tensor::to(Device device, bool non_blocking) const {
-  if (this->device() == device) {
-    return *this;
-  }
-  return detail::to(*this, options().device(device), non_blocking);
-}
 } // namespace at
 
 std::ostream& operator<<(
